intel_app/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiomysql==0.2.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
backoff==2.2.1
beautifulsoup4==4.13.0
bm25s==0.2.7.post1
boto3==1.36.11
botocore==1.36.11
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cobble==0.1.4
coloredlogs==15.0.1
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
dataclasses-json==0.6.7
deepsearch-glm==1.0.0
defusedxml==0.7.1
dill==0.3.9
distro==1.9.0
easyocr==1.7.2
emoji==2.14.1
et_xmlfile==2.0.0
eval_type_backport==0.2.2
faiss-cpu==1.10.0
fastapi==0.115.8
filelock==3.17.0
filetype==1.2.0
flatbuffers==25.1.24
fonttools==4.55.8
frozenlist==1.5.0
fsspec==2025.2.0
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-tools==1.70.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.28.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonlines==3.1.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
langchain==0.3.17
langchain-community==0.3.16
langchain-core==0.3.33
langchain-openai==0.3.3
langchain-text-splitters==0.3.5
langdetect==1.0.9
langsmith==0.3.4
latex2mathml==3.77.0
lazy_loader==0.4
lxml==5.3.0
mammoth==1.9.0
markdown-it-py==3.0.0
markdownify==0.14.1
markitdown==0.0.1a3
marko==2.1.2
MarkupSafe==3.0.2
marshmallow==3.26.0
matplotlib==3.10.0
mdurl==0.1.2
mmh3==4.1.0
mpire==2.10.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.17
mypy-extensions==1.0.0
mysql-connector-python==9.2.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.3
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
olefile==0.47
onnx==1.17.0
onnxruntime==1.20.1
openai
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pathvalidate==3.2.3
pdf2image==1.17.0
pdfminer.six==20240706
pi_heif==0.21.0
pillow==10.4.0
pinecone==5.4.2
pinecone-text==0.9.0
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
protoc-gen-openapiv2==0.0.1
psutil==6.1.1
puremagic==1.28
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
PyMuPDF==1.25.2
PyMySQL==1.1.1
pyparsing==3.2.1
pypdf==5.2.0
pypdfium2==4.30.1
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2025.1.28
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.1
python-pptx==1.0.2
pytz==2025.1
PyYAML==6.0.2
RapidFuzz==3.12.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
Rtree==1.3.0
s3transfer==0.11.2
safetensors==0.5.2
scikit-image==0.25.1
scikit-learn==1.6.1
scipy==1.15.1
semchunk==2.2.2
setuptools==80.9.0
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.14.1
SQLAlchemy==2.0.37
starlette==0.45.3
sympy==1.13.1
tabulate==0.9.0
tavily-python==0.5.0
tenacity==9.0.0
threadpoolctl==3.5.0
tifffile==2025.1.10
tiktoken==0.8.0
timm==1.0.14
together==1.3.14
tokenizers==0.21.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.48.2
triton==3.2.0
typer==0.12.5
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
unstructured==0.16.17
unstructured-client==0.29.0
unstructured-inference==0.8.6
unstructured.pytesseract==0.3.13
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
webencodings==0.5.1
wget==3.2
wrapt==1.17.2
XlsxWriter==3.2.2
yarl==1.18.3
youtube-transcript-api==0.6.3
zstandard==0.23.0
selenium==4.32.0
dbutils==3.1.1
mlflow==3.1.4
wget==3.2
mammoth==1.9.0
markdown==3.8.2
websocket-client==1.8.0
websockets==15.0.1
wheel
agno==2.1.5
unstructured-pytesseract